refactor(shortest_common_prefixes): Split into modules
Since the only difference is `pluck`, I split it into 2 modules, that then I can benchmark test
